+++ packages/compliance/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.1.2](https://github.com/RedHatInsights/javascript-clients/compare/@redhat-cloud-services/compliance-client-1.1.1...@redhat-cloud-services/compliance-client-1.1.2) (2024-11-18)
+
+### Dependency Updates
+
+* `@redhat-cloud-services/javascript-clients-shared` updated to version `1.2.6`
+## [1.1.1](https://github.com/RedHatInsights/javascript-clients/compare/@redhat-cloud-services/compliance-client-1.1.0...@redhat-cloud-services/compliance-client-1.1.1) (2024-11-18)
+
+### Dependency Updates
+
+* `@redhat-cloud-services/javascript-clients-shared` updated to version `1.2.5`
 ## [1.1.0](https://github.com/RedHatInsights/javascript-clients/compare/@redhat-cloud-services/compliance-client-1.0.0...@redhat-cloud-services/compliance-client-1.1.0) (2024-11-15)